+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 Changes:
+* Add persistent configs(closes #10)
 * CPU usage improved by removing unnecessary UI updates being done even without new messages